Fix audio format convertation

+++ audio_classification_service/modules/socket.py
@@ -8,7 +8,7 @@ async def socket_handler(request):
     await ws.prepare(request)
     async for ms in ws:
         message = json.loads(ms.data)
-        sample, sr = np.array(message["data"]), message["sr"]
+        sample, sr = np.array(message["data"], dtype=np.float32), message["sr"]
         prediction = request.app["model"](sample, sr)
         await ws.send_json({"prediction": prediction, "timestamp": message["timestamp"]})
     return ws
